+++ src/main/java/com/burgosmanuel/portfolio/educacion/IEducacionService.java
@@ -5,11 +5,11 @@ import java.util.List;
 public interface IEducacionService {
     public void agregarEducacion(Educacion edu);
     
-    public void editarEducacion(int id, Educacion datosEducacion);
+    public void editarEducacion(Long id, Educacion datosEducacion);
     
     public List<Educacion> listarEducacion();
     
-    public Educacion buscarEducacion(int id);
+    public Educacion buscarEducacion(Long id);
     
-    public void borrarEducacion(int id);
+    public void borrarEducacion(Long id);
 }
