+++ src/main/java/com/burgosmanuel/portfolio/proyecto/IProyectoService.java
@@ -6,11 +6,11 @@ import java.util.List;
 public interface IProyectoService {
     public void agregarProyecto(Proyecto proj);
     
-    public void editarProyecto(int id, Proyecto datosProyecto);
+    public void editarProyecto(Long id, Proyecto datosProyecto);
     
-    public Proyecto buscarProyecto(int id);
+    public Proyecto buscarProyecto(Long id);
     
     public List<Proyecto> listarProyectos();
     
-    public void borrarProyecto(int id);  
+    public void borrarProyecto(Long id);  
 }
